@@ -15,8 +15,8 @@ repository = "https://github.com/fedora-infra/fasjson"
 homepage = "https://github.com/fedora-infra/fasjson"
 
 [tool.poetry.dependencies]
-python = "^3.6.2"
-Flask = "^2.0.2"
+python = "^3.9.0"
+Flask = "^2.1.1"
 python-ldap = "^3.2.0"
 dnspython = "^2.1.0"
 flask-restx = "^0.5.0"
@@ -28,7 +28,7 @@ requests-gssapi = "^1.2.3"
 
 [tool.poetry.dev-dependencies]
 flake8 = "^4.0.1"
-pytest = "^6.2.5"
+pytest = "^7.1.2"
 pytest-mock = "^3.6.1"
 pytest-cov = "^3.0.0"
 liccheck = "^0.6.4"
@@ -36,7 +36,6 @@ bandit = "^1.6"
 black = "^22.0"
 isort = "^5.10.1"
 safety = "^1.10.3"
-typing = "*"
 pre-commit = "^2.16.0"
 
 [tool.black]
